+++ R/dispatcher.R
@@ -7,27 +7,20 @@
 #' controlled by [daemons()] and this function should not need to be called
 #' directly.
 #'
-#' The network topology is such that a dispatcher acts as a gateway between the
-#' host and daemons, ensuring that tasks received from the host are dispatched
-#' on a FIFO basis for processing. Tasks are queued at the dispatcher to ensure
-#' tasks are only sent to daemons that can begin immediate execution of the
-#' task.
+#' Dispatcher acts as a gateway between the host and daemons, dispatching tasks
+#' on a FIFO basis. Tasks are queued until a daemon is available for immediate
+#' execution.
 #'
-#' @inheritParams daemons
-#' @param host the character URL dispatcher should dial in to, typically an IPC
-#'   address.
-#' @param url the character URL dispatcher should listen at (and daemons should
-#'   dial in to), including the port to connect to e.g. tcp://hostname:5555' or
-#'   'tcp://10.75.32.70:5555'. Specify 'tls+tcp://' to use secure TLS
-#'   connections.
-#' @param n if specified, the integer number of daemons to be launched locally
-#'   by the host process.
+#' @param host (character) URL to dial into, typically an IPC address.
+#' @param url (character) URL to listen at for daemon connections, e.g.
+#'   'tcp://hostname:5555'. Use 'tls+tcp://' for secure TLS.
+#' @param n (integer) number of local daemons launched by host.
 #'
-#' @return Invisible NULL.
+#' @return Invisibly, an integer exit code: 0L for normal termination.
 #'
 #' @export
 #'
-dispatcher <- function(host, url = NULL, n = 0L, ...) {
+dispatcher <- function(host, url = NULL, n = 0L) {
   cv <- cv()
   sock <- socket("rep")
   on.exit(reap(sock))
@@ -61,8 +54,6 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
   serial <- res[[4L]]
   res <- res[[5L]]
 
-  inq <- outq <- list()
-  connections <- count <- 0L
   envir <- new.env(hash = FALSE, parent = emptyenv())
   `[[<-`(envir, "stream", res)
 
@@ -72,136 +63,11 @@ dispatcher <- function(host, url = NULL, n = 0L, ...) {
         cv_signal(cv) || wait(cv) || return()
       }
     }
-
-    changes <- read_monitor(m)
-    for (item in changes) {
-      item > 0 &&
-        {
-          connections <- connections + 1L
-          outq[[as.character(item)]] <- `[[<-`(
-            `[[<-`(
-              `[[<-`(`[[<-`(new.env(parent = emptyenv()), "pipe", item), "msgid", 0L),
-              "ctx",
-              NULL
-            ),
-            "sync",
-            FALSE
-          )
-          send(psock, list(next_stream(envir), serial), mode = 1L, block = TRUE, pipe = item)
-        }
-    }
   } else {
     listen(psock, url = url, tls = tls, fail = 2L)
-    url <- sub_real_port(psock, url)
+    url <- attr(attr(psock, "listener")[[1L]], "url")
   }
   send(sock, url, mode = 2L, block = TRUE)
 
-  ctx <- .context(sock)
-  req <- recv_aio(ctx, mode = 8L, cv = cv)
-  res <- recv_aio(psock, mode = 8L, cv = cv)
-
-  suspendInterrupts(
-    while (wait(cv)) {
-      changes <- read_monitor(m)
-      is.null(changes) ||
-        {
-          for (item in changes) {
-            if (item > 0) {
-              connections <- connections + 1L
-              outq[[as.character(item)]] <- `[[<-`(
-                `[[<-`(
-                  `[[<-`(`[[<-`(new.env(parent = emptyenv()), "pipe", item), "msgid", 0L),
-                  "ctx",
-                  NULL
-                ),
-                "sync",
-                FALSE
-              )
-              send(psock, list(next_stream(envir), serial), mode = 1L, block = TRUE, pipe = item)
-              cv_signal(cv)
-            } else {
-              id <- as.character(-item)
-              if (length(outq[[id]])) {
-                outq[[id]][["msgid"]] &&
-                  send(outq[[id]][["ctx"]], .connectionReset, mode = 1L, block = TRUE)
-                outq[[id]] <- NULL
-              }
-            }
-          }
-          next
-        }
-
-      if (!unresolved(req)) {
-        value <- .subset2(req, "value")
-
-        if (value[1L] == 0L) {
-          id <- readBin(value, integer(), n = 2L)[2L]
-          if (id == 0L) {
-            awaiting <- length(inq)
-            executing <- sum(as.logical(unlist(lapply(outq, .subset2, "msgid"), use.names = FALSE)))
-            found <- c(length(outq), connections, awaiting, executing, count - awaiting - executing)
-          } else {
-            found <- FALSE
-            for (item in outq) {
-              item[["msgid"]] == id &&
-                {
-                  send(psock, 0L, mode = 1L, pipe = item[["pipe"]], block = TRUE)
-                  found <- TRUE
-                  break
-                }
-            }
-            if (!found) {
-              for (i in seq_along(inq)) {
-                inq[[i]][["msgid"]] == id &&
-                  {
-                    inq[[i]] <- NULL
-                    found <- TRUE
-                    break
-                  }
-              }
-            }
-          }
-          send(ctx, found, mode = 2L, block = TRUE)
-        } else {
-          count <- count + 1L
-          msgid <- .read_header(value)
-          inq[[length(inq) + 1L]] <- list(ctx = ctx, req = value, msgid = msgid)
-        }
-        ctx <- .context(sock)
-        req <- recv_aio(ctx, mode = 8L, cv = cv)
-      } else if (!unresolved(res)) {
-        value <- .subset2(res, "value")
-        id <- as.character(pipe_id(res))
-        res <- recv_aio(psock, mode = 8L, cv = cv)
-        .read_marker(value) &&
-          {
-            send(outq[[id]][["ctx"]], value, mode = 2L, block = TRUE)
-            send(psock, 0L, mode = 2L, pipe = outq[[id]][["pipe"]], block = TRUE)
-            outq[[id]] <- NULL
-            next
-          }
-        send(outq[[id]][["ctx"]], value, mode = 2L, block = TRUE)
-        `[[<-`(outq[[id]], "msgid", 0L)
-      }
-
-      if (length(inq)) {
-        for (item in outq) {
-          item[["msgid"]] ||
-            {
-              if (.read_marker(inq[[1L]][["req"]])) {
-                item[["sync"]] && next
-                `[[<-`(item, "sync", TRUE)
-              } else if (item[["sync"]]) {
-                lapply(outq, `[[<-`, "sync", FALSE)
-              }
-              send(psock, inq[[1L]][["req"]], mode = 2L, pipe = item[["pipe"]], block = TRUE)
-              `[[<-`(item, "ctx", inq[[1L]][["ctx"]])
-              `[[<-`(item, "msgid", inq[[1L]][["msgid"]])
-              inq[[1L]] <- NULL
-              break
-            }
-        }
-      }
-    }
-  )
+  invisible(suspendInterrupts(.dispatcher(sock, psock, m, .connReset, serial, envir, next_stream)))
 }